Overhauled configuration mechanism
Modules now require a composite container factory and can create composite containers.
This functionality previously existed in `AbstractBaseModularModule` and has now been shifted to `AbstractBaseModule`.

Descendants can now create a composite container and return it from `setup()`.
This composite container can contain the config, by attaching two child containers:
- one having the service definitions
- the other having the config (such as the result of `_loadPhpConfigFile()`)

This allows modules to have configuration within their container, which is then attached to a composite container by
a parent modular module. Since the top-level composite container is then passed to each module's `run()` method and
as the argument for all service definitions, all config is made available via the container's `get()` method.
