+++ src/Module/AbstractBaseModularModule.php
@@ -2,13 +2,7 @@
 
 namespace RebelCode\Modular\Module;
 
-use ArrayAccess;
-use Dhii\Data\Container\ContainerFactoryInterface;
 use Dhii\Modular\Module\ModuleInterface;
-use Dhii\Util\String\StringableInterface as Stringable;
-use Psr\Container\ContainerInterface;
-use stdClass;
-use Traversable;
 
 /**
  * Base functionality for modular modules.it
@@ -25,37 +19,6 @@ abstract class AbstractBaseModularModule extends AbstractBaseModule
     use ModularModuleTrait;
 
     /**
-     * The factory to use for creating composite containers.
-     *
-     * @since [*next-version*]
-     *
-     * @var ContainerFactoryInterface
-     */
-    protected $compContainerFactory;
-
-    /**
-     * Initializes the modular module with all required information.
-     *
-     * @since [*next-version*]
-     *
-     * @param ContainerFactoryInterface                     $compContainerFactory The composite container factory.
-     * @param ContainerFactoryInterface                     $containerFactory     The container factory.
-     * @param string|Stringable                             $key                  The module key.
-     * @param string[]|Stringable[]                         $dependencies         The module dependencies.
-     * @param array|ArrayAccess|stdClass|ContainerInterface $config               The module config.
-     */
-    protected function _initModularModule(
-        ContainerFactoryInterface $compContainerFactory,
-        ContainerFactoryInterface $containerFactory,
-        $key,
-        $dependencies = [],
-        $config = []
-    ) {
-        $this->_setCompositeContainerFactory($compContainerFactory);
-        $this->_initModule($containerFactory, $key, $dependencies, $config);
-    }
-
-    /**
      * {@inheritdoc}
      *
      * @since [*next-version*]
@@ -63,43 +26,5 @@ abstract class AbstractBaseModularModule extends AbstractBaseModule
     protected function _getModuleServiceKey(ModuleInterface $module)
     {
         return sprintf('%s_module', $module->getKey());
-    }
-
-    /**
-     * Retrieves the composite container factory associated with this module.
-     *
-     * @since [*next-version*]
-     *
-     * @return ContainerFactoryInterface The composite container factory instance.
-     */
-    protected function _getCompositeContainerFactory()
-    {
-        return $this->compContainerFactory;
-    }
-
-    /**
-     * Sets the composite container factory for this module.
-     *
-     * @since [*next-version*]
-     *
-     * @param ContainerFactoryInterface $compContainerFactory The composite container factory instance.
-     */
-    protected function _setCompositeContainerFactory(ContainerFactoryInterface $compContainerFactory)
-    {
-        $this->compContainerFactory = $compContainerFactory;
-    }
-
-    /**
-     * Creates a composite container with the given children container instances.
-     *
-     * @since [*next-version*]
-     *
-     * @param ContainerInterface[]|Traversable $containers The children container instances.
-     *
-     * @return ContainerInterface The created composite container instance.
-     */
-    protected function _createCompositeContainer($containers)
-    {
-        return $this->_getCompositeContainerFactory()->make(['containers' => $containers]);
     }
 }
